knowledge-base-ai/backend/requirements.txt
# Web framework
fastapi==0.110.0
uvicorn[standard]==0.27.1
gunicorn==21.2.0

# Data validation & settings
pydantic==2.6.1
pydantic-settings==2.2.1

# Database (SQLAlchemy 2.0 style + Alembic migrations)
sqlalchemy==2.0.27
alembic==1.13.1
aiosqlite==0.20.0          # Async SQLite driver
asyncpg==0.29.0            # Async PostgreSQL driver (server deployment)

# Vector database
chromadb==0.5.3

# Embeddings (local model)
sentence-transformers==2.7.0
torch==2.2.1               # Required by sentence-transformers

# LLM client (Ollama HTTP API)
httpx==0.27.0

# Document parsing
pymupdf==1.24.0            # PDF parsing (PyMuPDF)
python-docx==1.1.0         # Word .docx parsing
markdown==3.6              # Markdown parsing
openpyxl==3.1.2            # Excel .xlsx parsing

# Structured logging
structlog==24.1.0

# Utilities
python-multipart==0.0.9    # For file upload support in FastAPI
python-dotenv==1.0.1       # .env file loading
